Sentiment_Analysis/sa_model/requirements.txt
Bootstrap-Flask==1.5.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
google==3.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
grpcio-tools==1.38.0
Jinja2==2.11.3
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
protobuf==3.16.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
sklearn==0.0
SQLAlchemy==1.4.7
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
threadpoolctl==2.2.0
Werkzeug==1.0.1
WTForms==2.3.3
